+++ backend/src/modules/context/services/ContextService.ts
@@ -5,8 +5,6 @@ import {inject, injectable} from 'inversify';
 import {ClientSession, ObjectId} from 'mongodb';
 import {IContext} from '#root/shared/interfaces/models.js';
 import {InternalServerError, BadRequestError} from 'routing-controllers';
-import {IQuestionRepository} from '#root/shared/database/interfaces/IQuestionRepository.js';
-import {dummyQuestions} from '../utils/questionGen.js';
 import { QuestionService } from '#root/modules/question/services/QuestionService.js';
 
 @injectable()
