+++ backend/src/modules/core/classes/transformers/Context.ts
@@ -1,28 +1,28 @@
-import {
-  ObjectIdToString,
-  StringToObjectId,
-} from '#shared/constants/transformerConstants.js';
-import { IContext } from '#shared/interfaces/models.js';
-import { Expose, Transform } from 'class-transformer';
-import { ObjectId } from 'mongodb';
+// import {
+//   ObjectIdToString,
+//   StringToObjectId,
+// } from '#shared/constants/transformerConstants.js';
+// import { IContext } from '#shared/interfaces/models.js';
+// import { Expose, Transform } from 'class-transformer';
+// import { ObjectId } from 'mongodb';
 
-class Context implements IContext {
-  @Transform(ObjectIdToString.transformer, { toPlainOnly: true })
-  @Transform(StringToObjectId.transformer, { toClassOnly: true })
-  @Expose()
-  _id?: string | ObjectId;
+// class Context implements IContext {
+//   @Transform(ObjectIdToString.transformer, { toPlainOnly: true })
+//   @Transform(StringToObjectId.transformer, { toClassOnly: true })
+//   @Expose()
+//   _id?: string | ObjectId;
 
-  @Expose()
-  text: string;
+//   @Expose()
+//   text: string;
 
-  @Expose()
-  createdAt?: Date;
+//   @Expose()
+//   createdAt?: Date;
 
-  constructor(data?: Partial<IContext>) {
-    // this._id = data?._id ? new ObjectId(data._id) : undefined;
-    this.text = data?.text;
-    this.createdAt = data?.createdAt ?? new Date();
-  }
-}
+//   constructor(data?: Partial<IContext>) {
+//     // this._id = data?._id ? new ObjectId(data._id) : undefined;
+//     this.text = data?.text;
+//     this.createdAt = data?.createdAt ?? new Date();
+//   }
+// }
 
-export { Context };
+// export { Context };
